Add state flag replacement
Using state flag 516 is better because 512 broke games that were fully up to date. 516 will let games that don't need an update to automatically become state flag 4, and games that do require an update will automatically become state flag 518. Neither of which happened if the state flag were to be 512.